* 'master' of github.com:tensorflow/tensorflow: (108 commits)
  Port person_detection example to Arduino
  Automated rollback of commit 05b15600b3c4472ec79aa865ea1d313c87b68a21
  Add support for some multi-store cases in affine fusion
  Internal cleanup: simplify the interface of converted_call, and shorten the names of function scopes. This in turn allows for much less verbose generated code.
  Lower ReluGrad to HLO.
  Log Graph to MLIR import failure in MLIRImportPass
  Add GraphDebugInfo to SavedModelBundle.
  [XLA] Increase the tolerable error for ErfInv(F16) tests.
  Instruct DynamicPadder to insert dynamic Unpad operations.
  [XLA:Python] Add support for infeed of tuple values.
  Automated rollback of commit 8f90210c1ee51e2bf893492f9dc0ee5f6d90fedf
  Introduce builtin_op_kernels.h header to simplify op registration
  [TFLite/MLIR] Adds tfl.transpose_conv op.
  Return error instead of failing on assert.
  Do not dump MLIR module on verifier failure.
  Remove dev-only log lines from user builds
  Add DVLOG macro to discard logs on user builds
  Fix bug where shape information that was supposed to be masked was visible inside the body of a tf.function for composite tensors such as RaggedTensor and SparseTensor.
  Update the usage and comments in define_inst.sh.
  Add exp operation to LLVMOPs.td.
  ...

+++ tensorflow/python/distribute/input_lib_test.py
@@ -524,7 +524,8 @@ class DistributedIteratorTensorTypeTest(DistributedIteratorTestBase,
     def dataset_fn(ctx=None):
       ctx = ctx or distribute_lib.InputContext()
       batch_size = ctx.get_per_replica_batch_size(global_batch_size)
-      row_lengths = np.mod(np.arange(20), 4)  # Deal with partial batches.
+      # Use 20 which isn't divisible by 8 to test partial batch behavior.
+      row_lengths = np.mod(np.arange(20), 4).astype(np.int64)
       ragged_tensor = ragged_tensor_lib.RaggedTensor.from_row_lengths(
           np.repeat(np.arange(20, dtype=np.float32), row_lengths), row_lengths)
       dataset = dataset_ops.DatasetV2.from_tensor_slices({
